go/runtime/bundle: Move metadata structs to component file
